Move types to config.ts and consequential type fixes.
